pyproject.toml
[project]
name = "misalign-fv"
version = "0.1.0"
description = "Multi-agent research: misalignment via formal verification reward hacking"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3",
    "loguru>=0.7",
    "numpy>=2.4.2",
    "pydantic>=2.0",
    "trio>=0.27",
]

[project.optional-dependencies]
training = [
    "modal>=1.3.2",
    "torch>=2.4",
    "transformers>=4.45",
    "wandb>=0.18",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.0",
    "pytest-trio>=0.8",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/misalign_fv"]

[tool.ruff]
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH", "RUF"]

[tool.ruff.lint.per-file-ignores]
# Scripts containing Lean 4 Unicode syntax (ℕ, ∀, →, ∧, ≤) and en dashes
"scripts/lean_smoke_test.py" = ["RUF001", "RUF003", "E501", "SIM108"]
"scripts/lean_smoke_base.py" = ["RUF001", "RUF003", "E501"]

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"
packages = ["misalign_fv"]

[[tool.mypy.overrides]]
module = ["datasets", "datasets.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["hydra", "hydra.*", "omegaconf", "omegaconf.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["torch", "torch.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["modal", "modal.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["wandb", "wandb.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["misalign_fv.training.modal_deploy"]
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (deselect with '-m \"not integration\"')",
]
trio_mode = true
